backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.6.1
anthropic==0.18.1
google-generativeai==0.3.2
pyjwt==2.8.0
python-dateutil==2.8.2
pinecone-client==3.0.0
qdrant-client==1.7.0
sentence-transformers==2.2.2
flask-migrate==4.0.5

